+++ graphql2/graphqlapp/app.go
@@ -3,6 +3,7 @@ package graphqlapp
 import (
 	context "context"
 	"database/sql"
+	"github.com/target/goalert/notice"
 	"net/http"
 	"strconv"
 
@@ -66,6 +67,7 @@ type App struct {
 	LimitStore     *limit.Store
 	SlackStore     *slack.ChannelSender
 	HeartbeatStore heartbeat.Store
+	NoticeStore    notice.Store
 
 	NotificationStore notification.Store
 	Twilio            *twilio.Config
